Upgrade Account authority from byte to short

+++ src/ChickenAPI.Packets/ServerPackets/Visibility/InCharacterSubPacket.cs
@@ -8,7 +8,7 @@ namespace ChickenAPI.Packets.ServerPackets.Visibility
     public class InCharacterSubPacket : PacketBase
     {
         [PacketIndex(0)]
-        public byte Authority { get; set; }
+        public short Authority { get; set; }
 
         [PacketIndex(1)]
         public GenderType Gender { get; set; }
